requirements.txt
mesa==2.1.2
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
stable-baselines3==2.0.0
torch==2.0.1
jupyter==1.0.0
pyyaml==6.0
scipy==1.10.1
scikit-learn==1.3.0
tqdm==4.65.0
